chore: update predefined

+++ src/modules/consultation/router.py
@@ -4,10 +4,12 @@ from typing import Annotated
 
 from fastapi import APIRouter
 
-from src.api.dependencies import DEPENDS_VERIFIED_REQUEST, DEPENDS_CONSULTATION_REPOSITORY
+from src.api.dependencies import DEPENDS_VERIFIED_REQUEST, DEPENDS_CONSULTATION_REPOSITORY, DEPENDS_USER_REPOSITORY
+from src.api.exceptions import ForbiddenException
 from src.modules.auth.schemas import VerificationResult
 from src.modules.consultation.repository import ConsultationRepository
 from src.modules.consultation.schemas import ViewConsultant, CreateConsultant, CreateTimeslot, AddAppointment
+from src.modules.user.repository import UserRepository
 
 router = APIRouter(prefix="/consultation", tags=["Consultation"])
 
@@ -15,8 +17,13 @@ router = APIRouter(prefix="/consultation", tags=["Consultation"])
 @router.post("/consultants/", status_code=201)
 async def post_consultant(
     data: CreateConsultant,
+    verification: Annotated[VerificationResult, DEPENDS_VERIFIED_REQUEST],
+    user_repository: Annotated[UserRepository, DEPENDS_USER_REPOSITORY],
     consultation_repository: Annotated["ConsultationRepository", DEPENDS_CONSULTATION_REPOSITORY],
 ) -> ViewConsultant:
+    user = await user_repository.read(verification.user_id)
+    if not user.is_admin:
+        raise ForbiddenException()
     return await consultation_repository.create_consultant(data)
 
 
@@ -38,8 +45,13 @@ async def get_consultant(
 async def add_timeslot(
     consultant_id: int,
     data: CreateTimeslot,
+    verification: Annotated[VerificationResult, DEPENDS_VERIFIED_REQUEST],
+    user_repository: Annotated[UserRepository, DEPENDS_USER_REPOSITORY],
     consultation_repository: Annotated["ConsultationRepository", DEPENDS_CONSULTATION_REPOSITORY],
 ) -> ViewConsultant:
+    user = await user_repository.read(verification.user_id)
+    if not user.is_admin:
+        raise ForbiddenException()
     return await consultation_repository.add_timeslot(consultant_id, data)
 
 
@@ -47,8 +59,13 @@ async def add_timeslot(
 async def remove_timeslot(
     consultant_id: int,
     timeslot_id: int,
+    verification: Annotated[VerificationResult, DEPENDS_VERIFIED_REQUEST],
+    user_repository: Annotated[UserRepository, DEPENDS_USER_REPOSITORY],
     consultation_repository: Annotated["ConsultationRepository", DEPENDS_CONSULTATION_REPOSITORY],
 ) -> ViewConsultant:
+    user = await user_repository.read(verification.user_id)
+    if not user.is_admin:
+        raise ForbiddenException()
     return await consultation_repository.remove_timeslot(consultant_id, timeslot_id)
